enh(#12): Sensor and actuator noise injection interface
Refactored the following sensors and actuators to have uniform XML/code
interfaces for noise injection:

  - epuck proximity sensor
  - eyebot light sensor
  - footbot base ground, motor ground, light sensor
  - generator battery, ground, light, positioning, proximity sensors
  - Colored blob camera and differential steering sensor
  - RAB sensor/actuator

Make documentation for the noise injector queryable so that it doesn't have to
be repeated everywhere.

Noise injection was removed from the footbot distance scanner because though it
was parsed from the XML tree, it was never used in the code.