enh(#12): Sensor and actuator noise injection interface
Refactored the following sensors and actuators to have uniform XML/code interfaces
for noise injection:

  - epuck proximity sensor
  - eyebot light sensor
  - footbot base ground, motor ground, light sensor
  - generator battery, ground, light, positioning, proximity sensors

These are simple sensors/actuators which had similar XML configuration which
different only in naming and whether or not the type of noise injected was
uniform or Gaussian.

Noise injection was removed from the footbot distance scanner because though it
was parsed from the XML tree, it was never used in the code.

The following sensors/actuators were not refactored, as they had more complex
noise injection models:

  - generic blob camera sensor
  - generic differential steering sensor/actuator